crawler/requirements.txt
scrapy>=2.11.0
psycopg2-binary>=2.9.9
itemadapter>=0.7.0
beautifulsoup4>=4.12.0
spacy>=3.7.0
requests>=2.31.0
